backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
dj-database-url
psycopg2-binary
gunicorn
python-dotenv
pandas
whitenoise
setuptools